+++ tests/faces/test_toroidalface3d.py
@@ -1,5 +1,7 @@
 import math
+import numpy as npy
 import unittest
+import volmdlr
 from volmdlr import faces, surfaces, wires
 
 
@@ -17,6 +19,25 @@ class TestToroidalFace3D(unittest.TestCase):
         face = faces.ToroidalFace3D.from_contours3d(surface, [contour])
         self.assertAlmostEqual(face.surface2d.area(), math.pi**2, 4)
         self.assertTrue(face.surface2d.outer_contour.is_ordered())
+
+    def test_planeface_intersections(self):
+        expected_results = [[14.700000000000001], [9.388571325840358], [9.282044389126776], [9.107655247892733],
+                            [8.870824266395537], [8.582455249249456], [5.000000000001248, 5.00000000000125],
+                            [3.717515834742253, 3.7176228274330256], [3.325526484035388, 3.426440202827559],
+                            [3.081959775756376, 3.081960004770588]]
+        ts = surfaces.ToroidalSurface3D(volmdlr.OXYZ, 2, 1)
+        tf = faces.ToroidalFace3D.from_surface_rectangular_cut(ts, -1.4, 3.5, 0., 2.5)
+
+        list_expected_lenghts1 = []
+        plane1 = surfaces.Plane3D(volmdlr.OXYZ)
+        plane1 = plane1.rotation(volmdlr.O3D, volmdlr.Z3D, math.pi / 4)
+        for i, n in enumerate(npy.linspace(0, math.pi / 4, 10)):
+            plane = plane1.rotation(plane1.frame.origin, volmdlr.X3D, n)
+            plane_face = faces.PlaneFace3D.from_surface_rectangular_cut(plane, 4, -4, 4, -4)
+            plane_intersections = tf.face_intersections(plane_face)
+            list_expected_lenghts1.append([i.length() for i in plane_intersections])
+            for result, expected_result in zip(plane_intersections, expected_results[i]):
+                self.assertAlmostEqual(result.length(), expected_result)
 
 
 if __name__ == '__main__':